+++ index.js
@@ -1,31 +1,9 @@
-import NotAllowedParameterError from './src/utils/shared/errors/NotAllowedParameter.error.js';
-import { initializeVariables, isObject, isString, objectInArrayContainsId } from './src/utils/shared/functions/validateTypes.js';
-import { create } from './src/main/dataSet/create.js';
+import { isObject, isString } from './src/utils/shared/functions/validateTypes.js';
 import IsNotObjectError from './src/utils/shared/errors/IsNotObject.error.js';
 import { doSearch } from './src/main/search/search.js';
 import IsNotStringError from './src/utils/shared/errors/IsNotString.error.js';
 import { getMaxMinDistance, resetMaxMin, setMaxMinDistance } from './src/main/search/findSimilarMatches.js';
-
-export let dataset = {
-    _array: [],
-    createdAt: undefined,
-    get array() {
-        return this._array;
-    },
-    set array(newArray) {
-        try {
-            if (!Array.isArray(newArray)) {
-                throw new NotAllowedParameterError('newArray', 'array');
-            }
-            this._array = newArray;
-        } catch (e) {
-            if (e.type && e.description)
-                console.error(`\n[${e.type}] - ${e.description}`);
-            else
-                console.error(e.message)
-        }
-    }
-}
+import { generate, findDataset } from './src/main/dataSet/generateDataset.js';
 
 /**
  * This function generates a data set based on an array of objects. **ID field required on each object**.
@@ -51,30 +29,6 @@ export function dataSetGenerate(main) {
     }
 }
 
-function generate(main) {
-    try {
-        initializeVariables(main);
-        dataset.array = [];
-
-        if (main.array.length > 0 && objectInArrayContainsId(main.array)) {
-            main.array.forEach(obj => {
-                if (!obj.hasOwnProperty(main.nameId)) throw new Error(`\'${main.nameId}\' does not match the object ID field name.`);
-
-                dataset.array.push(create(obj, main.nameId, main.wordSize, main.attributes));
-            })
-        }
-
-    } catch (e) {
-        if (e.type && e.description)
-            console.error(`\n[${e.type}] - ${e.description}`);
-        else
-            console.error(e.message)
-    }
-    dataset.createdAt = new Date();
-    return dataset.array;
-}
-
-
 /**
  * This function uses user input to search for objects generated by the data set.
  * - by default, it takes the first attribute that is considered a string in the object.
@@ -95,6 +49,13 @@ export function search(input, priorityAttribute) {
         console.error(`\n[${e.type}] - ${e.description}`);
 
     }
+}
+
+/**
+ * This function returns an object with all the attributes of the dataset.
+ */
+export function getDataset() {
+    return findDataset();
 }
 
 /**
