+++ readme.md
@@ -1,4 +1,4 @@
-# Data-search-js 1.2.3
+# **Data-search-js 1.2.5**
 
 <img src="./icon.png" width="250">
 
@@ -6,10 +6,11 @@
 
 This is a library that aims to perform a search for data in an array of objects. The parameter for such a search is a manual search for a user in a search bar.
 
-## Summary
+## **Summary**
 
 
   - [Dataset](#dataset)
+    - [Using getDataset](#using-getDataset)
   - [Search](#search)
     - [Priority attribute](#priority-attribute)
     - [Tips to use the priority attribute](#tip-to-use-the-priority-attribute)
@@ -36,7 +37,7 @@ This is a library that aims to perform a search for data in an array of objects.
 <br/>
 <br/>
 
-## Dataset
+## **Dataset**
 
 
 **The first step is to generate the data set.**
@@ -104,12 +105,57 @@ console.log(result)
 ]
 ```
 - The result will be an array of objects where, for each object in that array, an attribute called 'tags' will be added, which is an array with the filtered attribute values. 
-- These tags will be used by the search functionality.
+- \'totalSearchesFound\' will also be added.
+- Both attributes will be used as internal parameters by the search functionality.
+
+### **Using getDataset**
+- getDataset()
+- This function returns an object with all the attributes of the dataset.
+>`{`
+
+> **`_array : [{}]` // array with the generated objects.**.
+
+> **`createdAt : Date` // Data set creation date.**
+
+> **`array: [Getter/Setter]` // getters and setters for the array**
+
+>` }`
+
+<br/>
+
+> Code:
+```
+const info = getDataset(); 
+console.log(info);
+```
+> Output:
+```
+{
+  _array: [
+    {
+      id: 1,
+      title: 'Lord of the Rings',
+      genre: 'ADVENTURE',
+      totalSearchesFound: 0,
+      tags: [Array]
+    },
+    {
+      id: 2,
+      title: 'Fury',
+      genre: 'WAR',
+      totalSearchesFound: 0,
+      tags: [Array]
+    }
+  ],
+  createdAt: 2020-10-21T13:13:05.238Z,
+  array: [Getter/Setter]
+}
+```
 
 <br/>
 <br/>
 
-## Search
+## **Search**
 
 **The second step is to use the search function.**
 
@@ -138,7 +184,7 @@ console.log(result2);
 <br/>
 <br/>
 
-## Priority attribute
+## **Priority attribute**
 
 It is used when an object does not have enough information to be returned by the first parameter of the search function, but it needs to be returned, because the object has an attribute and this attribute has the necessary value for the return.
 
@@ -204,7 +250,7 @@ Even if the second object does not have all the information you typed in, it sti
 
 <br/>
 
-## Tip to use the priority attribute
+## **Tip to use the priority attribute**
 
 - The idea for the use of this priority attribute could be in the case of your application has search not only by field but also by filters, in the case of filters of genre of films (adventure, action etc...). 
 - So besides waiting for the user to type something, the application can also wait for him to pass a filter. 
@@ -216,7 +262,7 @@ search( res, 'genre');
 ```
 - The result would be the objects searched by typing but with priorities in the filters.
 
-## Search distance
+## **Search distance**
 
 - The search uses a distance system to find the results. 
 - This distance is delimited between two ends called 'min' and 'max'.
@@ -224,7 +270,7 @@ search( res, 'genre');
   - If nothing is found, the search uses the value between 'max' and 'min' to find the words (typed in the search) that most closely match the object's tag array.
   - At this point the closer the 'min' is to 'max' or the 'max' to 'min', the lower the tolerance for the words in the search and the greater the probability of finding an object similar (or not) to the one requested .
 
-### Using setDistance
+### **Using setDistance**
 - setDistance(min,max)
 - This function allows changing the maximum and minimum tolerance that the search needs to bring results.
 - This function accepts only fractional numeric values between 0 and 1.
@@ -278,7 +324,7 @@ console.log(search('fu'));
 - Now the search tolerance is only 0.3 (0.9-0.6), increasing the possibility of finding objects even with wrong words or in half.
 [RECOMMENDED] - For a more accurate search, i recommend using 'max' as being greater than 8, and working on 'min' with values between 2 to 7.5.
 
-### Using getDistance
+### **Using getDistance**
 - getDistance()
 - To find out what the 'min' and 'max' values are, use 'getdistance', this function returns an object with the updated 'min' and 'max' values:
 
@@ -291,7 +337,7 @@ console.log(getSearchDistance());
 { min: 0.3, max: 0.85 }
 ```
 
-### Using resetDistance
+### **Using resetDistance**
 - resetDistance()
 - Resets the 'min' and 'max' values to the default values.
 'min' = 0.3.
